+++ src/App.css
@@ -1,6 +1,3 @@
-
-
-
 /*------------ [_Root] -----------*/
 :root {
