+++ src/App.tsx
@@ -1,4 +1,8 @@
 import './App.css';
+import { GlobalStyleComponent } from 'styled-components';
+
+//The global style is supposed to be the general styling for all the components. 
+//Try applying it here and discard the App.css
 
 function App() {
   return (
